fix: corrected `type-limits` and `missing-field-initializers` warnings
When compiling a translation unit which includes the `nbnet.h` header as
well as its implementation (`#define NBNET_IMPL`) with all warnings
(`-Wall`) and extra warnings (`-Wextra`), it spits some warnings, most
of them being of the `type-limits` kind [1], and some more being of the
`missing-field-initializers` kind [2].

[1] Most of the warnings appear in places where comparisons involving
`unsigned` data types (e.g. `unsigned int`) and the number 0. For
instance (`@@ -3987,7 +3987,7 @@`):
```c
unsigned int id;
if (id < 0 || id >= NBN_RPC_MAX)
```
In this case, checking if the value stores within an `unsigned int` is less than 0 will always return false (0), as the `unsigned` keyword specifies that the corresponding variable can only represent non-negative values, so when storing a negative value inside it it will result in actually storing a 0. As a result of this, all these comparisons are unnecessary.

[2] In the diff `@@ -1731,10 +1731,10 @@`, it's being created an array
of `NBN_Driver` with a max capacity of `NBN_MAX_DRIVERS` and initialized
with 4 instances of the aforementioned `struct`. All 4 instances are
equal, effectively assigning -1 to the `id` field of the `NBN_Driver`
struct (which is an `int` type, so it's all good). It also assigns `NULL` to the second field, but it leaves the last field "uninitialized"; in quotes because it effectively zero initializes it. The edited version of this, which fixes the warning, is to assign to all 4 `struct`'s `{ .id = -1 }`, which only assigns -1 to the `id` field, leaving all the other fields zero initialized (in case of `char *` type, it's equivalent to assign to `NULL` or to 0).

Signed-off by: iWas-Coder <wasymatieh01@gmail.com>