fix: unnecessary pragmas and clang-specific attributes
The compiler pragmas surrounding the `nbn_drivers` array are not
needed anymore, due to the warnings were resolved in the previous
commit (90553e5).

Also, there are some attributes being defined for the `poly1305_auth`
function which are undefined for GCC (clang-specific), so an aditional
condition to check for that has been added.

Signed-off by: iWas-Coder <wasymatieh01@gmail.com>